Implement Client Certificate based Authentication
Also adds JSR-250 (@RolesAllowed and friends) Authorization annotations
on the Jersey methods.

Previously existing 'Auth' header authentication is still in place.
Users authenticating in this way are granted the 'JANITOR' role.

In reality the SSL Certificate security is handled by our proxy, which
denies users outright if they fail to present an SSL certificate with a
DN which we recognize. Requests with a valid, recognized DN are proxied
to Alexandria Webapp and the certificate DN is passed via HTTP header.

Requests authenticated with a Client Certificate via our proxy are granted
the CERTIFIED role.